Split allocators from subnet handling in pkg/network/master/subnets.go
